Added tests for the maxlength property.
The maxlength property needst to work with both UTF8 (or unicode)
and ASCII text. The class automatically determines which one to use
based on the extensions installed. This test sends in a Unicode
string that is longer than the maxlength of the name (64 characters)
and one that is exactly the length of the name after processing.