Added method drop(), refactored method save().
The save() method was refactored to be more accurate. It now attempts
to take the type of the data it is saving and bind it to an appropriate
PDO datatype. The Model now stores the type of data being saved.
For example, if a field is defined as NULL and can accept NULL values,
setting the value to a PHP NULL will result in the field being NULL
and not empty. It supports the different PHP PDO datatypes: INTEGER,
STRING, BOOL, and eventually LOB.

The new method drop() will take the tablename from a Model and drop
the table associated with it.